Change inputs and results from data magnitude converters to doubles instead of unsigned long long

+++ Blaze/io.cpp
@@ -34,7 +34,7 @@ inline unsigned long long		bytes_in_exa		= (long long)bytes_in_petta	* 1000u;
 inline unsigned long long		bytes_in_zetta		= bytes_in_exa				* 1000u;
 inline unsigned long long		bytes_in_yotta		= bytes_in_zetta			* 1000u;
 
-void convert_data_magnitude_to_bits(unsigned long long value, BYTE_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
+void convert_data_magnitude_to_bits(double value, BYTE_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
 {
 	try
 	{
@@ -272,7 +272,7 @@ void convert_data_magnitude_to_bytes(unsigned long long value, BIT_MAGNITUDE fro
 	}
 }
 
-void convert_data_magnitude_in_bits(unsigned long long value, BIT_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
+void convert_data_magnitude_in_bits(double value, BIT_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
 {
 	try
 	{
@@ -391,7 +391,7 @@ void convert_data_magnitude_in_bits(unsigned long long value, BIT_MAGNITUDE from
 	}
 }
 
-void convert_data_magnitude_in_bytes(unsigned long long value, BYTE_MAGNITUDE from_magnitude, BYTE_MAGNITUDE to_magnitude)
+void convert_data_magnitude_in_bytes(double value, BYTE_MAGNITUDE from_magnitude, BYTE_MAGNITUDE to_magnitude)
 {
 	try
 	{
@@ -510,7 +510,7 @@ void convert_data_magnitude_in_bytes(unsigned long long value, BYTE_MAGNITUDE fr
 	}
 }
 
-unsigned long long convert_data_magnitude_to_bits_copy(unsigned long long value, BYTE_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
+double convert_data_magnitude_to_bits_copy(double value, BYTE_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
 {
 	try
 	{
@@ -629,7 +629,7 @@ unsigned long long convert_data_magnitude_to_bits_copy(unsigned long long value,
 	}
 }
 
-unsigned long long convert_data_magnitude_to_bytes_copy(unsigned long long value, BIT_MAGNITUDE from_magnitude, BYTE_MAGNITUDE to_magnitude)
+double convert_data_magnitude_to_bytes_copy(double value, BIT_MAGNITUDE from_magnitude, BYTE_MAGNITUDE to_magnitude)
 {
 	try
 	{
@@ -748,7 +748,7 @@ unsigned long long convert_data_magnitude_to_bytes_copy(unsigned long long value
 	}
 }
 
-unsigned long long convert_data_magnitude_in_bits_copy(unsigned long long value, BIT_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
+double convert_data_magnitude_in_bits_copy(double value, BIT_MAGNITUDE from_magnitude, BIT_MAGNITUDE to_magnitude)
 {
 	try
 	{
@@ -867,7 +867,7 @@ unsigned long long convert_data_magnitude_in_bits_copy(unsigned long long value,
 	}
 }
 
-unsigned long long convert_data_magnitude_in_bytes_copy(unsigned long long value, BYTE_MAGNITUDE from_magnitude, BYTE_MAGNITUDE to_magnitude)
+double convert_data_magnitude_in_bytes_copy(double value, BYTE_MAGNITUDE from_magnitude, BYTE_MAGNITUDE to_magnitude)
 {
 	try
 	{
